streamex v0.3.0 Streamex.Client

The Streamex.Client module is not meant to be used directly, although request-related functions are publicly available in case of need.

Functions

execute_request(req)

Executes a Streamex.Request. Returns {:ok, json}, or {:error, message} if something went wrong.

prepare_request(req)

Prepares a Streamex.Request for execution. Basically finalizes a request url. Returns a Streamex.Request.

sign_request(req)

Signs a Streamex.Request. If token info are attached to the request, it will be signed with a JWT Token. Otherwise, with a server signature. Returns a Streamex.Request.