Streamex v0.3.0 Streamex.Token

The Streamex.Token module is responsible for JWT Token generation. It’s not meant to be used directly, although functions are publicly available in case of need.

Types

t :: %Streamex.Token{action: term, feed_id: term, resource: term}

Functions

compact(token, secret)

Generates a compact token, with grants for given Streamex.Token’s’ resource, action and feed_id, and signs it with secret.